requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-Limiter==3.3.0
Flask-CORS==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pytz==2022.7.1
cx-Oracle==8.3.0
six==1.16.0
Werkzeug==2.2.3
python-dotenv==1.0.0
requests==2.28.2
cryptography==39.0.1
gunicorn==20.1.0
flasgger==0.9.5
apispec==6.3.0
marshmallow==3.20.1